+++ src/modules/LectureHall/components/AllocatedLH.js
@@ -1,14 +1,23 @@
 import Lecturehall from "./Lecturehall";
 import { useState, useEffect, useRef } from "react";
+import Modal from "./Modal";
+import Backdrop from "./Backdrop";
+import "./Lecturehall.css";
 
 export default function AllocatedLH() {
   const [allocatedLH, setAllocatedLH] = useState([1, 5, 7]);
+  const [modalIsOpen, setModalIsOpen] = useState(false);
   const [toDel, setToDel] = useState();
   const didMount = useRef(false);
 
   function deleteHandler(id) {
     setAllocatedLH(allocatedLH.filter((lh) => lh !== id));
     setToDel(id);
+    setModalIsOpen(false);
+  }
+
+  function setModalHandler() {
+    setModalIsOpen(false);
   }
 
   useEffect(() => {
@@ -25,16 +34,20 @@ export default function AllocatedLH() {
 
   return (
     <div className="flex flex-col items-center justify-center min-h-screen">
-      <h1 className="py-5">Lecture Halls Allocated to you:</h1>
+      <h1 className="py-5">Lecture Halls Allocated to you:-</h1>
       {allocatedLH.map((lh) => (
-        <div key={lh}>
+        <div key={lh} className='card'>
           <Lecturehall id={lh} />
-          {/* <button
-            className="p-2 border-2 rounded text-red border-red hover:text-red-400 mb-2"
-            onClick={() => deleteHandler(lh)}
+          <div className='actions'>
+          <button
+            className="btn"
+            onClick={() => setModalIsOpen(true)}
           >
             Remove
-          </button> */}
+          </button>
+          </div>
+          {modalIsOpen && <Modal onCancel={setModalHandler} onConfirm={()=>deleteHandler(lh)}/>}
+          {modalIsOpen && <Backdrop onClick={setModalHandler}/>}
         </div>
       ))}
     </div>
